Optimize sqlalchemy inserts
