+++ tests/formats/sql/test_adapter.py
@@ -1,5 +1,4 @@
 import pytest
-import sqlite3
 import datetime
 import sqlalchemy as sa
 from frictionless import Package, Resource, formats
@@ -254,8 +253,7 @@ def test_sql_adapter_dialect_basepath_issue_964(sqlite_url):
         ]
 
 
-# TODO: recover
-@pytest.mark.skip
+@pytest.mark.ci
 def test_sql_adapter_max_parameters_issue_1196(sqlite_url, sqlite_max_variable_number):
     # SQLite applies limits for the max. number of characters in prepared
     # parameterized SQL statements, see https://www.sqlite.org/limits.html.
@@ -284,28 +282,3 @@ def test_sql_adapter_max_parameters_issue_1196(sqlite_url, sqlite_max_variable_n
         resource.write(
             sqlite_url, control=formats.SqlControl(table="test_max_param_table")
         )
-
-
-# Fixtures
-
-
-@pytest.fixture
-def sqlite_max_variable_number():
-    # Return SQLite max. variable number limit, set as compile option, or
-    # default.
-    #
-    # Default value for stock SQLite >= 3.32.0
-    # (https://www.sqlite.org/limits.html#max_variable_number): 32766
-    #
-    # Note that distributions *do* customize this e.g. Ubuntu 20.04:
-    # MAX_VARIABLE_NUMBER=250000
-    conn = sqlite3.connect(":memory:")
-    try:
-        with conn:
-            result = conn.execute("pragma compile_options;").fetchall()
-    finally:
-        conn.close()
-    for item in result:  # type: ignore
-        if item[0].startswith("MAX_VARIABLE_NUMBER="):
-            return int(item[0].split("=")[-1])
-    return 32766
